feat: Enables tests with custom participant management.
